Remove render of template from unknown exception
Rendering a template for an error outside of the route results in
Dancer2 throwing its own warning.  The on_route_exception hook used
to be the place to coerce the exception into a specific response.
At some point this changed and no longer works as I expected.

Instead, we're just going to log the exception to make sure we're
aware internally, then let Dancer2 handle rendering the 500.tt
view on it's own.  This app doesn't have any information to
propagate to the user here anyway.