Standardize config file location and name
This commit standardizes the config file for the project to be
loaded from a more generally expected place.

The config was previously loaded from the project root as a dot rc
file.  Now, the config file will be loaded from either:

- $HOME/.config/pasteburn/config.ini
- /etc/pasteburn/config.ini

The tests were reworked to support this change and are included in
this commit.