Add runmode new to session
This commit moves the runmode query parameter into the session
to display the flash message for success.  The rm query parameter
was a cheap workaround to easily display the success message.

More thought and work should go into get and set flash methods,
or if flash messages are even needed for the functionality at all.
For now, this change retains current functionality in the ui.