Update crypt hash to bcrypt
This commit replaces Crypt::PBKDF2 with
Crypt::Eksblowfish::Bcrypt.

Although Crypt::PBKDF2 has an easier interface, HMACSHA1 with so
few iterations is not as secure as bcrypt.  Raising the
interactions to a generally accepted secure amount (although still
HMACSHA1 instead of HMACSHA512), is slower by an order of
magnitude not (IMO) acceptable for a production application.