Add copy to clipboard on secret link
This commit updates the secret link to copy the URL to clipboard
when clicked.  For now, the secret link doesn't degrade gracefully
if JS is disabled (to be added later).

An external dependency, clipboard.js, was added.  More work to
update the UI will be done, at which time clipboard.js may not be
needed.