Add button to view secret as author
Previously, the secret template doesn't allow the author to view
the secret they just created.  Looking at web server logs, there
are a fair number of users who create a secret but never decrypt
it.  From a UX perspective, I suspect not being able to test
viewing their own secret leads them to believe the application
doesn't work; then they don't come back to use the service again.

To view your secret as the author, you either have to clear the
cookie for the site, or view the secret link in a different browser
that doesn't contain the cookie (this was by design).

To allow users to test their own secret, this commit adds a view
button to the secret view.  When the button is clicked, the secret
id is removed from the cookie and the user redirected to view the
secret as they would have been if they weren't the author.