Add configurable option for blank passphrase
This commit adds a configurable option to allow users to create
secrets with a blank passphrase.

 [passphrase]
 allow_blank = 0

To allow blank passphrases to be submitted, the allow_blank value
can be set to 1.

If set to 0 (dont allow blank passphrases), the frontend post route
will verify both passphrase and secret inputs contain values.
Empty string is not allowed for passphrase.

If set to 1 (allow blank passphrases) the frontend post route will
only verify the secret input contains a value.  The passphrase
input will then submit an empty string to the backend.

Changes to the backend were made to allow empty string for
passphrase to accommodate this change.