Remove default_duration and expires age settings
Previously the project set default expire times for the session,
storage object, and secret hard coded to 7 days.

Since the secret age can now be configured to a value other than 7
days, to allow all of these settings to stay in sync, this commit
updates the app config and Storage modules to no longer set expire
times.

Documentation for both Dancer2::Session::Cookie and
Session::Storage::Secure explain not setting a value means no
expiration time is enforced.  Since secrets are expired based on
the created_at field in the row, and not according to the cookie
or storage object expiration, this change keeps all places within
the project in line for expiration dates.