Update html font-family to generic web fonts
Loading non-generic web fonts come with a higher risk of browser
behavior changing the font behavior if fonts are not kept up to
date.  In this case, we were previously using the defaults set by
the Skeleton CSS library, which admittedly is very out of date.

In a newer version of Chrome, these non-standard fonts were
behaving strangely, displaying italic in places not defined in the
CSS.  Updating to the generic web fonts makes the sections behave
better without sacrificing too much consistency between browsers.