Update create_test_app to validate config params
The create_test_app method bypasses the config validation, which
means the frontend integration tests won't validate the application
in a way similar to how it would run in production.

This change changes the config loading to override the config
module's _load instead of get, which means the config being passed
in will now be verified for all required params and values.